Разделение на модули: main.js data.js util.js
